+++ libminifi/src/core/Processor.cpp
@@ -41,129 +41,6 @@ using namespace std::literals::chrono_literals;
 
 namespace org::apache::nifi::minifi::core {
 
-ProcessorMetrics::ProcessorMetrics(const Processor& source_processor)
-    : source_processor_(source_processor) {
-  on_trigger_runtimes_.reserve(STORED_ON_TRIGGER_RUNTIME_COUNT);
-}
-
-std::string ProcessorMetrics::getName() const {
-  return source_processor_.getProcessorType() + "Metrics";
-}
-
-std::unordered_map<std::string, std::string> ProcessorMetrics::getCommonLabels() const {
-  return {{"metric_class", getName()}, {"processor_name", source_processor_.getName()}, {"processor_uuid", source_processor_.getUUIDStr()}};
-}
-
-std::vector<state::response::SerializedResponseNode> ProcessorMetrics::serialize() {
-  std::vector<state::response::SerializedResponseNode> resp;
-
-  state::response::SerializedResponseNode root_node;
-  root_node.name = source_processor_.getUUIDStr();
-
-  state::response::SerializedResponseNode iter;
-  iter.name = "OnTriggerInvocations";
-  iter.value = static_cast<uint32_t>(iterations.load());
-
-  root_node.children.push_back(iter);
-
-  state::response::SerializedResponseNode average_ontrigger_runtime_node;
-  average_ontrigger_runtime_node.name = "AverageOnTriggerRunTime";
-  average_ontrigger_runtime_node.value = static_cast<uint64_t>(getAverageOnTriggerRuntime().count());
-
-  root_node.children.push_back(average_ontrigger_runtime_node);
-
-  state::response::SerializedResponseNode last_ontrigger_runtime_node;
-  last_ontrigger_runtime_node.name = "LastOnTriggerRunTime";
-  last_ontrigger_runtime_node.value = static_cast<uint64_t>(getLastOnTriggerRuntime().count());
-
-  root_node.children.push_back(last_ontrigger_runtime_node);
-
-  state::response::SerializedResponseNode transferred_flow_files_node;
-  transferred_flow_files_node.name = "TransferredFlowFiles";
-  transferred_flow_files_node.value = static_cast<uint32_t>(transferred_flow_files.load());
-
-  root_node.children.push_back(transferred_flow_files_node);
-
-  for (const auto& [relationship, count] : transferred_relationships_) {
-    state::response::SerializedResponseNode transferred_to_relationship_node;
-    transferred_to_relationship_node.name = std::string("TransferredTo").append(1, toupper(relationship[0])).append(relationship.substr(1));
-    transferred_to_relationship_node.value = static_cast<uint32_t>(count);
-
-    root_node.children.push_back(transferred_to_relationship_node);
-  }
-
-  state::response::SerializedResponseNode transferred_bytes_node;
-  transferred_bytes_node.name = "TransferredBytes";
-  transferred_bytes_node.value = transferred_bytes.load();
-
-  root_node.children.push_back(transferred_bytes_node);
-
-  resp.push_back(root_node);
-
-  return resp;
-}
-
-std::vector<state::PublishedMetric> ProcessorMetrics::calculateMetrics() {
-  std::vector<state::PublishedMetric> metrics = {
-    {"onTrigger_invocations", static_cast<double>(iterations.load()), getCommonLabels()},
-    {"average_onTrigger_runtime_milliseconds", static_cast<double>(getAverageOnTriggerRuntime().count()), getCommonLabels()},
-    {"last_onTrigger_runtime_milliseconds", static_cast<double>(getLastOnTriggerRuntime().count()), getCommonLabels()},
-    {"transferred_flow_files", static_cast<double>(transferred_flow_files.load()), getCommonLabels()},
-    {"transferred_bytes", static_cast<double>(transferred_bytes.load()), getCommonLabels()}
-  };
-
-  {
-    std::lock_guard<std::mutex> lock(relationship_mutex_);
-    for (const auto& [relationship, count] : transferred_relationships_) {
-      metrics.push_back({"transferred_to_" + relationship, static_cast<double>(count),
-        {{"metric_class", getName()}, {"processor_name", source_processor_.getName()}, {"processor_uuid", source_processor_.getUUIDStr()}}});
-    }
-  }
-
-  return metrics;
-}
-
-void ProcessorMetrics::incrementRelationshipTransferCount(const std::string& relationship) {
-  std::lock_guard<std::mutex> lock(relationship_mutex_);
-  ++transferred_relationships_[relationship];
-}
-
-std::chrono::milliseconds ProcessorMetrics::getAverageOnTriggerRuntime() const {
-  if (on_trigger_runtimes_.empty()) {
-    return 0ms;
-  }
-  std::chrono::milliseconds sum = 0ms;
-  std::lock_guard<std::mutex> lock(average_on_trigger_runtime_mutex_);
-  for (const auto& runtime : on_trigger_runtimes_) {
-    sum += runtime;
-  }
-  return sum / on_trigger_runtimes_.size();
-}
-
-void ProcessorMetrics::addLastOnTriggerRuntime(std::chrono::milliseconds runtime) {
-  std::lock_guard<std::mutex> lock(average_on_trigger_runtime_mutex_);
-  if (on_trigger_runtimes_.size() < STORED_ON_TRIGGER_RUNTIME_COUNT) {
-    on_trigger_runtimes_.push_back(runtime);
-  } else {
-    if (next_average_index_ >= on_trigger_runtimes_.size()) {
-      next_average_index_ = 0;
-    }
-    on_trigger_runtimes_[next_average_index_] = runtime;
-    ++next_average_index_;
-  }
-}
-
-std::chrono::milliseconds ProcessorMetrics::getLastOnTriggerRuntime() const {
-  std::lock_guard<std::mutex> lock(average_on_trigger_runtime_mutex_);
-  if (on_trigger_runtimes_.empty()) {
-    return 0ms;
-  } else if (on_trigger_runtimes_.size() < STORED_ON_TRIGGER_RUNTIME_COUNT) {
-    return on_trigger_runtimes_[on_trigger_runtimes_.size() - 1];
-  } else {
-    return on_trigger_runtimes_[next_average_index_ - 1];
-  }
-}
-
 Processor::Processor(const std::string& name, std::shared_ptr<ProcessorMetrics> metrics)
     : Connectable(name),
       logger_(logging::LoggerFactory<Processor>::getLogger()),
