+++ libminifi/test/unit/ContentRepositoryDependentTests.h
@@ -27,6 +27,7 @@
 #include "core/Resource.h"
 #include "../TestBase.h"
 #include "../Catch.h"
+#include "../DummyProcessor.h"
 #include "StreamPipe.h"
 
 #pragma once
@@ -52,22 +53,6 @@ struct ReadUntilItCan {
     }
   }
 };
-
-class DummyProcessor : public core::Processor {
- public:
-  using core::Processor::Processor;
-
-  static constexpr const char* Description = "A processor that does nothing.";
-  static auto properties() { return std::array<core::Property, 0>{}; }
-  static auto relationships() { return std::array<core::Relationship, 0>{}; }
-  static constexpr bool SupportsDynamicProperties = false;
-  static constexpr bool SupportsDynamicRelationships = false;
-  static constexpr core::annotation::Input InputRequirement = core::annotation::Input::INPUT_ALLOWED;
-  static constexpr bool IsSingleThreaded = false;
-  ADD_COMMON_VIRTUAL_FUNCTIONS_FOR_PROCESSORS
-};
-
-REGISTER_RESOURCE(DummyProcessor, Processor);
 
 class Fixture {
  public:
